Frontend/api_server/requirements.txt
fastapi==0.95.1
pydantic==1.10.2
uvicorn
duckdb
httpx 
python-dotenv